+++ src/ska_low_mccs_pasd/field_station/field_station_device.py
@@ -97,8 +97,8 @@ class MccsFieldStation(MccsBaseDevice):
         }
         self._health_report = ""
         self._health_rollup = self._setup_health_rollup()
-        self.set_change_event("antennaPowerStates", True, self.VerifyEvents)
-        self.set_change_event("outsideTemperature", True, self.VerifyEvents)
+        self.set_change_event("antennaPowerStates", True, False)
+        self.set_change_event("outsideTemperature", True, False)
         self.init_completed()
 
         message = (